Updated for-loops to allow re-assignment of iterators
Previously, iterators for for-loops could either be pre-declared, in which case, they could not be re-initialized in the for-loop header, or they could be declared inline, in which case they HAD to be initialized in the header.

For-loops have been updated to make initialization optional when using a pre-declared iterator. This means that for-loops can now declare and initialize a new variable as an iterator, use a pre-declared variable as an iterator, or re-initialize a pre-declared variable as an iterator (with the latter being added with this change).

Parser tests and analyzer tests have been added and updated to account for this change.